requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2==2.9.9
drf-spectacular==0.27.2